Ensure repository set to writable after moving storage
When moving a repository to new storage it moves through multiple
states. When the repository move is replicated in the new storage we set
the repository to writable. However, if we receive an error, for example
a connection error, it is possible the repository-move state will be
updated while the repo-writeable state didn't change.

This change adds another before_transition callback to ensure that the
repo is set to writable if the replication transition failed in this
way.

Fixes: https://gitlab.com/gitlab-org/gitlab/-/issues/427254