Add VersionsFinder; update catalog resource and version models
Adds Ci::Catalog::Resources::VersionsFinder that will be used
in the future instead of ReleasesFinder. This new Finder is
similar to ReleasesFinder except that it returns Versions instead
of Releases. This MR also updates the catalog resource and
version models accordingly to use the new Versions association
instead of Releases.