Cleanup EnsureStringDetail Rubocop TODO
While I respect the attempt to DRY here, and even agree with it to some
extent, Rubocop _does not_ and we can't upset Rubocop, now can we?  Since
other API files repeat these strings, it seems fine to do so here as
well, which allows us to drop an entire category of Rubocop TODOs!