Set the height of the iframe containing credit card form to 480px
To accommodate Google reCAPTCHA challenge form, we set the height of the
iframe to 480px. Without doing this, the iframe height is insufficient
to display the challenge form in its entirety and the submit button of
the form is not visible to the user and cannot be clicked.

Changelog: fixed
EE: true