Serve pages deployments without going through PagesMetadatum
With the work to introduce [Gitlab Pages Multiple Versions], a project
now have to support multiple `PagesDeployment` at the same time.

For this reason, instead of using
`ProjectPagesMetadatum#pages_deployment_id` to identify valid deployment
the `PagesDeployment#deleted_at` field was [introduced].

For now, we're still serving a single pages version, to avoid changing
the current behavior, but instead of looking it up through
`ProjectPagesMetadatum#pages_deployment_id`, we're using the
`project.pages_deployments.main_version` to identify a project active
pages deployment.

This is an extraction of the [MR that introduce the multiple version
serving]. It was decided to not serve multiple versions on this commit
to avoid extra complexity on this, already, large change.

[Gitlab Pages Multiple Versions]: https://gitlab.com/groups/gitlab-org/-/epics/10914
[introduce]: https://gitlab.com/gitlab-org/gitlab/-/merge_requests/130848
[MR that introduce the multiple version serving]: https://gitlab.com/gitlab-org/gitlab/-/merge_requests/133261

Related to: https://gitlab.com/gitlab-org/gitlab/-/issues/416494

Changelog: changed