Do not use find_by inside a scope
find_by doesn't return an ActiveRecord::Relation so when it eventually
returns nil (i.e. when it can't find a record) the enclosing scope
returns Users::PhoneNumberValidation.all (an ActiveRecord::Relation)
instead of nil as expected. This results to the early return to be
skipped and the call to record.user throws
"undefined method `user' for #<ActiveRecord::Relation" error.

Changelog: fixed
EE: true