Add settings to allow merge without pipeline
Split out of
https://gitlab.com/gitlab-org/gitlab/-/merge_requests/135276, as it was
getting too large and we should start getting parts of this merged. This
only adds a feature flagged project setting (it does not implement the
updated mergeability check).

When "Pipelines must succeed" is selected, it is currently not possible
to merge without a pipeline. This includes cases where the configuration
is technically valid, but the pipeline is intentionally not created, such
as:

- pipeline was omitted because no jobs matched `rules` and `only/except`
- pipeline was omitted via `workflow:rules`
- there was no `.gitlab-ci.yml`

Changing this behavior is not backwards-compatible, because there is no
way of knowing if an external CI system is running for a given change.
So we add a project setting instead.

See https://gitlab.com/gitlab-org/gitlab/-/issues/334281

Changelog: changed