Extract more db/docs logic to ::Gitlab::Database::Dictionary
Recently we added the `Gitlab::Database::Dictionary` class. Initially it
only had the responsibilit of wrapping the content of a single
`db/docs/*.yml` file. But this MR moves all the logic in this clas down
to `Gitlab::Database::Dictionary::Entry` and adds a class methods
`Dictionary.entries` and `Dictionary.entry` to load a single entry. The
changes to this file will be most easily reviewed ignoring whitespace
because a lot of it has just been indented under `Entry`.

This refactoring moved the previous logic for building an array of
entries from the `GitlabSchema` class because that was not a suitable
place for generic handling of `db/docs/*.yml` files.

This will help with
https://gitlab.com/gitlab-org/gitlab/-/merge_requests/136682 to allow us
to reuse the dictionary building logic in validating the `sharding_key`.