Create more generic fetchUserCounts
The previous iteration of `refreshUserMergeRequestCounts` had a promise
based API which wasn't necessary and the name suggested that it only
updated a User's MR counts while under the hood it updated _all_ the
counts.

The function is now co-located with the super sidebar and has been
renamed `fetchUserCounts` which is more generic. It is in a separate
file to avoid side-effects when importing, which is ironic, because the
whole function triggers a side-effect.

All the uses of the function have been adjusted as well and tests for
them added.