Introduce `extended_trailers` field in commits API
Adds a new field that correctly supports all commit trailers. Unlike
the existing `trailers` field, this will return an array of values
where the same key is used multiple times, e.g. "Cc".

This functionality will eventually be moved over to the `trailers`
field once the original functionality has been deprecated. This
is a breaking change to the API and therefore can't be performed
in this commit.

Changelog: added