Add `PagesDeployment` upload_finished field
Before [serving pages directly from pages_deployments], we were creating
an `PagesDeployment` and in the commit callback, the ZIP file was
uploaded. Only after the `PagesDeployment` entry was committed to the
database, we would update the `ProjectPagesMetadatum` to indicate the
deployment was ready.

Now, [serving pages directly from pages_deployments], created a problem
where we're starting to serve a `PagesDeployment` before its upload is
ready.

To fix that, a new `upload_ready` column will be introduced, to
signalize when the new `PagesDeployment` upload is finished and we
start to serve it.

[serving pages directly from pages_deployments]: https://gitlab.com/gitlab-org/gitlab/-/merge_requests/135993

Changelog: fixed