Stop validating snippets when restoring from backup
Previously snippets have been validated when restoring a backup. This
causes confusing error messages and failure of snippets that work
(despite being invalid for having an incorrect branch name).

Restore should not be performing any data validation. It should instead
try and restore the system as close to how it was when the backup was
taken. This is a fundamental difference between importers and restoring
from backups.

Changelog: fixed