Ingest vulnerability from multiple pipelines
Continuous Vulnerability Scanning detects vulnerabilities across the
entire GitLab instance. As a performance improvement, we need to be able
to create batches of vulnerabilities which may be on multiple different
projects.

This change updates vulnerability ingestion to handle multiple pipelines
by storing them in the `finding_map` object.

Some tasks need to query for objects on the project and compare them
with objects found during ingestion. Since this won't scale for multiple
projects, a separate `IngestCvsSliceService` is defined which skips
these tasks.