Fix AiMessage attributes on Duo Chat QA tests
On https://gitlab.com/gitlab-org/gitlab/-/merge_requests/136354 we
changed the format of the AiMessage attributes so that the `resource`
was part of the `context`. In that MR we adjusted all the places where
an AiMessage is formed, except here where it slipped throuh, and went
unnoticed, since Chat QA tests are optional.