Fix token for repository tree keyset pagination
Contributes to https://gitlab.com/gitlab-org/gitlab/-/issues/377591

**Problem**

The page token generation was done on Rails side before. It used a last
element of the collection as a token.

It works for branches and tags, but the tree impementation requires to
provide a token from `GetTreeEntriesResponse`.

**Solution**

Fetch the cursor from the `GetTreeEntriesResponse` and correctly assign
it to Links header.

Refactor branches and tags finders to have a similar interface as
TreeFinder.

Changelog: fixed