ADD controller for ActivityPub subscription to releases actor
This is the third and last part of the merging the overarching
subscription MR at
https://gitlab.com/gitlab-org/gitlab/-/merge_requests/132460 .

This provides the controller, route and service used to create
subscriptions.

We add the `inbox` endpoint as specified by ActivityPub on our
`releases` actor to receive various post requests. The only requests we
accept are Follow and unfollow (Undo > Follow) ones. If we receive an
other kind of activity (that could be, for example, that someone
commented on our release post in the Fediverse), we silently discard it.

The service's role is to create the blank subscription and to queue the
background job that will resolve the various URLs and prepare the
subscription.