Sanitize string provided to to_tsvector
We sanitize full text queries for non-allowed characters, but
we don't do this for generating the search vector itself.

When we supply non-allowed characters to to_tsvector, it
can affect the resulting search vector.

Example: providing "<gitlab>" string will result in an empty
tsvector.

By sanitising the non-allowed characters out, it allows for
words surrounded by these non-allowed characters to be
included in the search vector