Remove gitlab:cleanup:moved rake task
This rake task was supposed to have been removed 4 years ago in
https://gitlab.com/gitlab-org/gitlab/-/issues/24633 but somehow got
missed.

This rake task tries to access gitaly controlled filesystems directly.
This does not work with gitaly cluster or remote gitalys. So should not
be supported.

While there is a spec called spec/gitlab/cleanup_rake_spec.rb but this
is not the test for this rake task.

Changelog: removed