Only "Edit status" in user bar once modal is mounted
Apparently there could be a race condition where the "Edit Status"
button didn't trigger the modal, if clicked too early. This mainly
manifested in the specs.

The old workaround was adding a "ready" class to some DOM element. We
now properly listen for a "mounted" event on the "Set Status Modal" and
only show it once it has been successfully mounted.