Ensure we provide strings to "Set Status Modal"
Apparently if some values are `null`, the modal had trouble with them.
Before this issue was masked because all the data was serialzed into
data attributes and then read as strings again.

This way, we set all falsy values to be the empty string instead and the
modal works fine again.