Require gitaly_address instead of path in storage settings
Requiring path in this context is a bit confusing because gitlab-rails
does not actually determine where repositories are stored, gitaly does.
Omnibus tends to hide this fact by using the same settings to configure
gitaly and to configure gitlab-rails.

So here we begin to untangle this configuration by ensuring that
gitlab-rails has the configuration that it needs.

Changelog: changed