Remove the repository downloads and repository data conflict check
This checks that the repository_downloads_path is not the same as any
configured storages. The problem is that gitlab-rails soon will not have
access to the repository storage paths. So this check will be
impossible.

Since there are already documented warnings and the archive cleaner
specifically only removes archives (not git repos). It should be
relatively safe, if a bit confusing, when misconfigured.

Changelog: removed