Add CI remote include parallelization
In preload_content, we start executing the HTTP call in another thread.
Then, when the content method is called, we wait for the HTTP response.

In the next step, we'll work on limiting the max number of parallel
HTTP calls in a single user call.

This change is behind the FF ci_parallel_remote_includes.