Drop Go 1.19 for workhorse
We moved to Go 1.20 in GitLab 16.2 and no GitLab versions in our
support policy is built using Go 1.19.

We are working on a feature based on the http.ResponseController type
that was introduced in Go 1.20 and to avoid porting this type to Go 1.19
we should simply drop the test suite.