feat: Disabled agents are not shown in the dropdown on the UI
This change makes sure that users are not able to select a agent
that is disabled from the UI when creating a new workspace.

Additionally, this change also ensures that when a workspace is
created via the API, it is not saved if the agent is not enabled.

Changelog: changed
EE: true