Fix offsets used for requesting last commit data
Tree entries for submodules are rendered before blobs but blobs are
earlier in the pagination.  This difference lead to an absence of
submodule data with blob data being requested instead.

This change is to make requests to the logs_tree endpoint while
taking into account this difference.

Contributes to: https://gitlab.com/gitlab-org/gitlab/-/issues/415711
Changelog: fixed