Create `organization_details` table to store organization description
`description_html` and `cached_markdown_version` will be used by
CacheMarkdownField for caching.

The `description` and the related columns are not stored in the
`organizations` table to avoid making it too wide which might
have an impact on its performance in the future. We're doing
the same with the `namespaces` table i.e. moving the description
related fields from the `namespaces` to the`namespace_details` table.