Add auto_cancel_on_new_commit to ci_pipeline_metadata and remove a const
This is the first step of supporting workflow:auto_cancel:on_new_commit.
In this change, we are adding a new column auto_cancel_on_new_commit to
the ci_pipeline_metadata table. Also, we are removing the NOT NULL
constraint of the name column because now a ci_pipeline_metadata row
can be created without a name.

Changelog: added