Add status to get container repository(ies) REST API response
The status attribute of container repositories is already included
in the response of the ContainerRepositoryDetails GraphQL API, but
they are not part of the equivalent REST APIs. This status
attribute is currently used to determine the status of the async
deletion for a given repository. The lack of this attribute in the
REST API has been identified as a problem for some users in
https://gitlab.com/gitlab-org/gitlab/-/issues/427032.

This change adds the `status` field to the response for
GET /registry/repositories/:id and
GET /projects/:id/registry/repositories endpoints.

Changelog: changed