Fix loading of ancestors for signed out users
When loading the ancestors for a work item, we optimize the querying of
the data by preloading resources based on the users' access.

We also tried to preload resources for signed out users, where
`current_user` was `nil`. This led to errors as the pre-loaders assume
to have a user set.

Changelog: fixed