Build QA image on stable branch pipelines (in canonical)
Part of https://gitlab.com/gitlab-com/gl-infra/delivery/-/issues/2825

As part of the release environments epic https://gitlab.com/groups/gitlab-com/gl-infra/-/epics/837
we have release environments running the version of GitLab for each stable
branch we support (currently last three stable branches).

In order to validate that any fixes developers merge to stable branches
not only deploys correctly, but passes QA, we want to run QA against
the release-environments after they have been deployed to successfully.

The first step for this, is to build the QA image needed to run the tests
(and the test code of the specific version of GitLab) on the stable branch
pipelines themselves. This way they can be consumed later in a job that
runs QA against the release-environments (to be implemented later).