Use new getTopFrequentItems helper
This updates the `CreateIssueForm` component to get frequently visited
projects with the newer `getTopFrequentItems` helper. The old helper
lives in a directory that will be removed as part of the new nav
cleanup.
The result should be essentially the same, except that we'll always show
a maximum of 3 projects regardless of the viewport's size.

Changelog: changed
EE: true