Search for users when focus on input
When inviting members to a group or project, when focusing on the input
field it searches for users.

Apparently, previously there some short of autofocus in the search input
field. In that situation searching for users automatically was
undesired. However, currently, we don't see any autofocus on the search
input field, so we enable the search on focus.

Changelog: changed