Add cost factored stats GraphQL query
The query will be used when the feature flag:
`display_cost_factored_storage_size_on_project_pages` is enabled.

The query is the same as the project_storage query, except it uses the
cost factored stats and it aliases all storage statistics type to the
same names as the original statistics object, so the frontend can just
use a different object name for cost factored statistics and all other
internal would still work the same way.