Fix adding confidential child tasks
We removed querying the `confidential` attribute for work items. As a
result users were no longer able to add existing confidential tasks to
confidential issues.

The reason was that we check in the frontend if the existing child task
is confidential, when the parent is as well. As we no longer queried
`confidential` in our GraphQL query, we assumed that the task was not
confidential and showed an error. The user was not able to add the task
as we disabled the submit button.

Changelog: fixed