Add generated_files check to Gitlab::Git::Compare
Change filter_generated_files as we actually just want to list
generated files instead. Also, it doesn't need to be at Repository
model so move it to the raw repository.

We also want to be able to detect generated files automatically,
but we need to wait for a new gitaly RPC for that so just hande manual
overrides for now.