Add `interruptible` to trigger jobs
When implementing new auto-cancel strategies, we realized that trigger
jobs don't accept `interruptible` and they have no effect on pipeline
auto-cancellation.

In this change, we are copying the `interruptible` keyword from builds
to triggers. Also, we are adding test cases for pipeline
auto-cancellation to show that trigger jobs do not affect
auto-cancellation.

Our main goal is that when we implement the "interruptible" auto-cancel
strategy, users can also set trigger jobs interruptible or not.

Changelog: added