Improves performance of runners list with large number of jobs
This change aims to improve the performances of runners lists
which contain large number of jobs. As each job count increases the
time it takes to load a runner, we split this field in a separate
request.

Additionally, this change batches queries by a "batchKey", so the client
can opt-in to handle the job count queries in batches to limit the
number of outgoing API requests from the browser.

Changelog: changed