Ensure packager ordering is consistent
It's common for there to be only 1-3 packagers in the dependency list.
This means that when sorting by packager, the ordering is essentially
random because all of the rows have the same value for it. This change
adds composite sorting on (packager, name) to ensure that the ordering
remains consistent.