Preserve page count in Catalog
When navigating to a details page of a catalog resource,
the page count was lost because the Vue app would unmount
and the pageCount was kept in the data attribute. Now,
we put the pageCount inside the Apollo cache so it can
stay in sync with the rest of the application.