Simplify initialisation of Whats New App
We don't need to trigger the whats new app from the legacy top nav any
longer. This allows us to simplify the init code in the super sidebar.

While we are at it, we are sneaking in a little rename of
`whats-new-app` to `js-whats-new-app`. Also, we are now passing the
version digest from the super sidebar, rather than reading it from the
DOM attributes.