Bump LocalStorageCache key version
When using stale data from localStorage for the query, we
caused a frontend console error in Apollo which caused the
import to fail to start.

By bumping the version key, we can clear all previous queries /
data and use the new version (that includes `hasFailures` field)
which does not cause errors.

Changelog: fixed