Implement background syncing for catalog_resources and projects
Adds catalog_resource_sync_events table and model
to serve as a queue of catalog resources that need to be synchronized
with data changes to their associated `projects` records. PG triggers
add SyncEvent records when a catalog resource is created or
when a relevant column in an associated project is updated.
A worker processes the sync events.

Changelog: added