Correctly deduplicate vulnerabilities with 8 digit image tags
Exclude 8-digit numbers from the location fingerprint when performing
vulnerability deduplication. This means that when two different images
are tagged with short hashes such as my-image:62011677 and
my-image:e2e32c98, these will be grouped in the vulnerability report
instead of being displayed as separate line items. It's still ambiguous
as to whether 62011677 is a hash or a number, but since version numbers
rarely become that large this is the quickest way to fix the case where
we have a short-ref tagging scheme.

Changelog: fixed
EE: true