Fix broken dependency list aggregation
7bd1e41b3a918a7749d301c509c9846868bf474a added an ignore_column
statement onto Sbom::Occurrences which causes ActiveRecord to use
`select(*model.column_names)` as the default select statement instead of
`SELECT *`. Since the `occurrence_count` and `project_count` columns are
defined inline using an `AS` statement, we were dependent on `SELECT *`
to have these columns appear in the result set. This introduced a
regression which caused the dependency list to not correctly group
projects and locations. To fix the regression, we explicitly select
these columns in order to ensure that they appear in the result set
while ignored columns do not.

Changelog: fixed
EE: true