Add spec for sharding key not null
As part of our
[Organization Isolation](https://docs.gitlab.com/ee/architecture/blueprints/organization/isolation.html)
work we are ensuring that all tables have a valid "sharding key". This
key will ultimately be used to efficiently identify which organization a
row belongs to.

In https://gitlab.com/gitlab-org/gitlab/-/merge_requests/136682 we added
a spec to ensure that all newly created tables have a sharding key. Now
we want to validate that these sharding keys are also not nullable.

Our spec needs to handle the following ways that we ensure columns are
not nullable:

1. Column is not nullable in the schema like `namespace_id NOT NULL`
2. Column is validated using a check constraint like
   `CONSTRAINT check_77fea3f0e7 CHECK ((namespace_id IS NOT NULL))`

This also adds a few more sharding keys to tables that are using a check
constraint rather than a `NOT NULL` column. This is to illustrate that
the spec does indeed handle both cases.