Remove legacy fetching for CI components
Originally, CI components could be defined anywhere in a resource
repository. Now, we require them to be inside a `templates` directory.
We supported the legacy fetching for a few milestones to give users time
to upgrade, and now we are removing it.

Related to https://gitlab.com/gitlab-org/gitlab/-/issues/415855
Changelog: removed