Update model to use new ignore_column remove_never attribute
Formerly we needed to set 'remove_after' to year 3000 as a kind of hack
in order for the 'ignore_column' helper to effectively ignore the
column forever. Now a new 'remove_never' attribute has been introduced
to the 'ignore_column' helper. So we are updating the model
Ci::Catalog::Resources::SyncEvent to use the new attribute.