Simplify page_with_sidebar_class helper
As the super sidebar is now always active, we can simplify the
`page_with_sidebar_class` helper quite a bit. It also allows us to
remove some other helpers:
- `ApplicationHelper#collapsed_sidebar?`
- `NavigationHelper#page_has_collapsed_sidebar?`
- `NavigationHelper#page_has_collapsed_super_sidebar?`
- `NavigationHelper#page_has_sidebar?`

The latter is always true on pages that render the new sidebar, because
we actually set it in: `app/views/layouts/_page.html.haml` if the super
sidebar is enabled.