Fix deprovisioning paid features upon transfer to subgroups
Prior to this change, when transferring a subgroup of a group with a
paid subscription to a group without a subscription, some paid features
weren't correctly deprovisioned.

This change fixes the deprovisioning of such a transfer by reusing the
deprovisioning logic of paid features in the `Projects::TransferService`
for the `Groups::TransferService`. Additionally, it also adds more tests
for the deprovisioning to have better test coverage.

Changelog: fixed
EE: true