Add MergeTrains::CreateRefService
The EE-only bit of `MergeRequests::CreateRefService` was actually
specific to merge trains and needed to be extracted before we could
re-use this service in other parts of the code. The idea is that
`MergeRequests::CreateRefService` should supersede
`MergeRequests::MergeToRefService`, in some (or all) cases. For example,
for merged result pipelines when automatic rebases are enabled.