Prefer it_behaves_like over include_examples
This changes the docs example for the use of 'an idempotent worker'
shared example to suggest `it_behaves_like` rather than
`include_examples`.

`it_behaves_like` creates a context around the shared examples, whereas
`include_examples` doesn't.

As the example group defines a `subject`, using `include_examples` leads
to the `subject` of the shared examples becoming to `subject` in the
context outside of the shared examples too, which results in a worker
that has its `#perform` method being called multiple times.

https://gitlab.com/gitlab-org/gitlab/-/blob/33b0faaae9e1d4fd907fda6c251a453ffc469b84/spec/support/shared_examples/workers/idempotency_shared_examples.rb#L22-28