Make events root property prohibited in redis_event metrics
The root property `events` is used by metrics with the `data_source: internal_events`.
Sometimes we forget to update the `data_source` when we migrate a metric from RedisHLL to InternalEvents.

This change makes sure the spec will fail if we add a `events` root property and keep the `redis_hll` `data_source`.