Do not fail import on NoMethodError
An import should be resilient to any non-stage worker hitting an error
(of any kind) and we shouldn't fail the import within our non-stage
workers.

This change switches the handling of a `NomethodError` to track the
error as normal. It is not re-raised because there's no point in the
worker retrying.

The motivation for this change was a particular situation where a
`NoMethodError` was being raised due to calling a method on a `nil`.

https://gitlab.com/gitlab-org/manage/import-and-integrate/discussions/-/issues/61#note_1689481972

Changelog: fixed