Move release-environments child pipeline to be sourced from master
Part of https://gitlab.com/gitlab-com/gl-infra/delivery/-/issues/19905

In order to make development of release-environments easier by reducing
backports and skew of code across stable branches, we want to make it
so regardless of GitLab version being deployed to a release-environment,
we are always using the child-pipeline setup from master.

This means we can can make changes to the pipeline setup on `master`,
and it will take effect across all stable branches.