Exclude each_batch from Style/SymbolProc rubocop rule
Prior to this change, the `Style/SymbolProc` rubocop rule reported
offenses if `each_batch` or `each_sub_batch` are written with a block
instead of `&:method`. Those two methods don't work with `&:method`
which makes is impossible to fix this offense without disabling the rule
inline.

This change excludes `each_batch` and `each_sub_batch` from the
`Style/SymbolProc` rubocop rule to address this issue. Additionally, it
is ensured that the existing allowed methods are added back to the
custom exclusion rule as well.

Existing inline comments to disable this rule are removed as a result of
this exclusion and the to do list for the rule is also updated to remove
files that only have an offense for these specific methods.

Changelog: changed