Include `blob_path` in `VulnerabilityLocationType`
The `VulnerabilityLocationType` graphql type is supposed to contain a
blob_path ([docs][0]), however this field is always `nil` because we
do not pass that data to location in the resolver.

The REST api was also returning a `nil` `blob_path`. This was because
the presenter operated on `string` keys, and the REST entity passes in
`symbol` keys. This issue was solved by making `location` have
indifferent access in the presenter

[0]:https://docs.gitlab.com/ee/api/graphql/reference/#vulnerabilitylocationsecretdetection
related to: https://gitlab.com/gitlab-org/gitlab/-/issues/434358

EE: true
Changelog: fixed