Add paths to deploy key, ensure return controller
For the frontend, add a number of paths that are required on the
frontend. These are edit, enable and disable.

Also, rails seems to return nothing instead of an empty list when there
are no deploy keys. The easiest way to fix this is to wrap the return
within a map.

Let's also add frontend fixtures for the new controller endpoints.