Fix metrics server not shutting down when Puma is stopped
Previously if a separate metrics server were configured and
Puma were stopped, the metrics server would continue to run. When
Puma is started again, it attempts to spawn a new metrics server,
but this fails repeatedly because the port has already been taken.

To fix this, add an `at_exit` handler that will shutdown the process
supervisor if Puma exists. This won't help the case if Puma is
abruptly killed, however.

Relates to https://gitlab.com/gitlab-org/omnibus-gitlab/-/issues/8109

Changelog: fixed