Move deprovisioning of paid group and project features
Prior to this change, the deprovisioning of paid group and project
features happened within the transaction to update the group or project.
This put a heavy load on the transaction and also made the
deprovisioning of pipeline subscriptions buggy. The bug was caused by
relying on a triggerer (group or project) to schedule the deprovisioning
with a `run_after_commit` hook.

This change moves the deprovisioning out of the transaction and triggers
it after the transaction has completed. The bug when deprovisioning is
removed as part of this move by removing the `run_after_commit` hook.

Changelog: fixed
EE: true