What's new app: Create own "mount element"
The What's new app is a singleton. Rather than creating an element which
we search for in the Vue App, then pass to the What's New App
initializer, we just create our own div. This probably can be refactored
in favor of something more Vue-y but for now it seems like a good
compromise.