Fix upload file dialog button
This tweaks a change made in
https://gitlab.com/gitlab-org/gitlab/-/merge_requests/137537
(77d7fd6c43f3b7380a8020242b681c835f0dc750).

Rather than finding the first `[data-button-type="attach-file"]` in the
document, find the first one in the form we've been given. If it doesn't
exist, then just make the textarea itself clickable.

This also completely removes the `handleAttachFile` method, as it didn't
seem to be contributing to the behaviour at all.